Separate Gateway interfaces from implementations
