+++ src/gateways/BasicAuthGateway.ts
@@ -1,12 +1,6 @@
 import { IncomingMessage, ServerResponse } from "http";
 import basicAuthMiddleware from "nextjs-basic-auth-middleware";
-
-export interface IBasicAuthGateway {
-  authenticate: (
-    request: IncomingMessage,
-    response: ServerResponse
-  ) => Promise<void>;
-}
+import { IBasicAuthGateway } from "./interfaces/IBasicAuthGateway";
 
 export class BasicAuthGateway implements IBasicAuthGateway {
   async authenticate(
