+++ src/gateways/interfaces/AuthGateway.ts
@@ -1,0 +1,5 @@
+import { ConfidentialClientApplication } from "@azure/msal-node";
+
+export interface AuthGateway {
+  getAccessToken: (cca?: ConfidentialClientApplication) => Promise<string>;
+}
