+++ src/gateways/interfaces/EmailServiceGateway.ts
