+++ src/gateways/interfaces/IBasicAuthGateway.ts
@@ -1,0 +1,8 @@
+import { IncomingMessage, ServerResponse } from "http";
+
+export interface IBasicAuthGateway {
+  authenticate: (
+    request: IncomingMessage,
+    response: ServerResponse
+  ) => Promise<void>;
+}
