+++ src/gateways/interfaces/UserSessionGateway.ts
@@ -1,5 +1,5 @@
 import { GetServerSidePropsContext } from "next";
-import { Session } from "./NextAuthUserSessionGateway";
+import { Session } from "../NextAuthUserSessionGateway";
 
 export interface UserSessionGateway {
   getSession: (context: GetServerSidePropsContext) => Promise<Session | null>;
