+++ src/lib/IAppContainer.ts
@@ -1,12 +1,12 @@
 import { IncomingMessage } from "http";
 import { DraftRegistration } from "../entities/DraftRegistration";
-import { AuthGateway } from "../gateways/AadAuthGateway";
-import { AccountHolderGateway } from "../gateways/AccountHolderGateway";
-import { IBasicAuthGateway } from "../gateways/BasicAuthGateway";
-import { BeaconGateway } from "../gateways/BeaconsApiBeaconGateway";
-import { DraftRegistrationGateway } from "../gateways/DraftRegistrationGateway";
-import { EmailServiceGateway } from "../gateways/EmailServiceGateway";
-import { UserSessionGateway } from "../gateways/UserSessionGateway";
+import { AccountHolderGateway } from "../gateways/interfaces/AccountHolderGateway";
+import { AuthGateway } from "../gateways/interfaces/AuthGateway";
+import { BeaconGateway } from "../gateways/interfaces/BeaconGateway";
+import { DraftRegistrationGateway } from "../gateways/interfaces/DraftRegistrationGateway";
+import { EmailServiceGateway } from "../gateways/interfaces/EmailServiceGateway";
+import { IBasicAuthGateway } from "../gateways/interfaces/IBasicAuthGateway";
+import { UserSessionGateway } from "../gateways/interfaces/UserSessionGateway";
 import { AddNewUseToDraftRegistrationFn } from "../useCases/addNewUseToDraftRegistration";
 import { AuthenticateUserFn } from "../useCases/authenticateUser";
 import { ClearCachedRegistrationFn } from "../useCases/clearCachedRegistration";
